Split editor console UI out of toolbar and table-drive action routing
Move console focus/selection/copy/scroll handling into
game_data_editor_console_ui.c so the toolbar module keeps to hit
classification and action routing. Replace the tool-mode if/else chain
with an editor_tool_mode_def table (special-case reset hooks stay
explicit) and the action-to-signal prefix chain with name/prefix
tables.
